+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2020-10-05
+
 ### Added
 
 - `icon`, `iconWidth` and `iconHeight` props to the interfaces `store-list` and `store-map`